initial commit of entitypicker
This is version 0.2.0 and depends on jQuery (I've tested with as low as
1.7.3 - can't imagine it requires anything more than when .on() was
introduced) and jQuery UI for the autocomplete control, styling and
effects. Supports multiple pickers on a page, different service
endpoints for each picker, form submission with input name pulled from
either the id of the container or an HTML5 style data attribute, adding
entities through javascript, querying out entities through javascript
and more, I'm sure. Lots of customization to get UI to be unobtrusive.