Added focus style for Issue #4
In webkit browsers an outline is placed around focused input elements. It wasn't ideal for the entity picker to have the 'cloaked' input be called out that way so it was disabled in the css. This change adds event listeners for focusin and focusout on the autocomplete input and applies the style to the parent picker instead. This makes the pickers behave consistently with other form fields in webkit browsers.