Add setting for max entities message
The entity picker sets a title attribute if there is the maxEntities
are >= 0. This was a simple message which specified the maximum
entities that could be selected. With this new version you can specify
a string or function (that received the current maxEntities as a
parameter) to set this message. The title attribute will be set if the
message string has a length > 0.