Various bug fixes, refactoring
Some bug fixes result in different functionality so I want to call this
a possible breaking change. This fixes broken support for local data by
adding delay setting(same usage as jquery autocomplete) as well as
calling autocomplete search on focus of the field if the field contains
the minimum number of characters to begin searching. This means clicking
a field bound to local data with a delay of 0 will display the local
data options for picking. This has a side effect of starting a search on
traditional fields if they happen to contain enough text but had lost
focus (began a search, clicked out / missed clicking the options then
clicked back into the search field).