Added changeSource function
Added new method to allow the changing of the autocomplete datasource.
This doesn't change any of the other settings, just the source. It's
most useful if your picker is using local data and you need to update
the options - for instance if your options change as a results of a user
picking a new option in a dropdown.